% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classDefinition.R
\docType{class}
\name{pixmap-class}
\alias{pixmap-class}
\alias{pixmap}
\title{define a virtual super class pixmap----}
\value{
return value
}
\description{
define a virtual super class pixmap----
}
\section{Slots}{

\describe{
\item{\code{size}}{integer.  size for the map}

\item{\code{cellres}}{numeric.  cellres in the map.}

\item{\code{bbox}}{numeric. bbox in}
}}

\examples{
pixmap();
}
