% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classDefinition.R
\docType{class}
\name{pixmapIndexed-class}
\alias{pixmapIndexed-class}
\alias{pixmapIndexed}
\title{define subclass pixmapIndexed}
\value{
return value.
}
\description{
define subclass pixmapIndexed
}
\section{Slots}{

\describe{
\item{\code{index}}{matrix. index}

\item{\code{col}}{character. col}
}}

\examples{
pixmapIndexed();
}
