% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classDefinition.R
\docType{class}
\name{pixmapRGB-class}
\alias{pixmapRGB-class}
\alias{pixmapRGB}
\title{define a subclass pixmapRGB}
\value{
return value
}
\description{
define a subclass pixmapRGB
}
\section{Slots}{

\describe{
\item{\code{red}}{matrix. red channel setting by matrix.}

\item{\code{green}}{matrix. green channel setting by matrix}

\item{\code{blue}}{matrix. blue channel setting by matrix}
}}

\examples{
pixmapRGB();
}
