% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,pixmapIndexed-method}
\alias{plot,pixmapIndexed-method}
\title{为pixmapIndexed定义plot方法}
\usage{
\S4method{plot}{pixmapIndexed}(x, y, xlab = "", ylab = "", axes = FALSE,
  asp = 1, ...)
}
\arguments{
\item{pixmapIndexed}{define pixmapIndexed object}
}
\value{
return value
}
\description{
为pixmapIndexed定义plot方法
}
\examples{
plot();
}
