% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,pixmapRGB-method}
\alias{plot,pixmapRGB-method}
\title{定义pixmapRGB的plot方法}
\usage{
\S4method{plot}{pixmapRGB}(x, y, ...)
}
\arguments{
\item{pixmapRGB}{pixmapRGB的对象}
}
\value{
返回值
}
\description{
定义pixmapRGB的plot方法
}
\examples{
plot();
}
