% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show,pixmap-method}
\alias{show,pixmap-method}
\title{为了pixmap定义show方法}
\usage{
\S4method{show}{pixmap}(object)
}
\arguments{
\item{pixmap}{定义一个pixmap对象}
}
\value{
返回值
}
\description{
为了pixmap定义show方法
}
\examples{
show();
}
